Move aggregate ID-type to its own file
