+++ src/shared/types/index.ts
@@ -1,2 +1,3 @@
+export * from './aggregate-ids.types.js';
 export * from './do-base.js';
 export * from './flavor.types.js';
